backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=2.1.0
python-multipart>=0.0.9
PyMuPDF>=1.24.0
google-genai>=0.8.0
httpx>=0.27.0
google-cloud-storage>=2.17.0
python-dotenv>=1.0.1
Pillow>=10.3.0
aiofiles>=23.2.1
manim>=0.18.0
sarvamai>=0.1.0
